+++ term/src/terminalstate/keyboard.rs
@@ -35,6 +35,7 @@ impl TerminalState {
                 encoding,
                 newline_mode: self.newline_mode,
                 application_cursor_keys: self.application_cursor_keys,
+                modify_other_keys: self.modify_other_keys,
             },
             is_down,
         )?;